+++ integrations/jina/src/haystack_integrations/components/embedders/jina/document_embedder.py
@@ -49,7 +49,7 @@ class JinaDocumentEmbedder:
         task: str | None = None,
         dimensions: int | None = None,
         late_chunking: bool | None = None,
-    ):
+    ) -> None:
         """
         Create a JinaDocumentEmbedder component.