+++ integrations/nvidia/src/haystack_integrations/utils/nvidia/models.py
@@ -35,7 +35,7 @@ class Model:
     def __hash__(self) -> int:
         return hash(self.id)
 
-    def validate(self):
+    def validate(self) -> int:
         if self.client:
             client = self.client if isinstance(self.client, Client) else Client.from_str(self.client)
             supported = {