+++ integrations/opensearch/src/haystack_integrations/components/retrievers/opensearch/sql_retriever.py
@@ -28,7 +28,7 @@ class OpenSearchSQLRetriever:
         document_store: OpenSearchDocumentStore,
         raise_on_failure: bool = True,
         fetch_size: int | None = None,
-    ):
+    ) -> None:
         """
         Creates the OpenSearchSQLRetriever component.